Added Webscraping Project
A skill optimization for the mobile game sword and glory where the character can select from a list of limited skills that become more available through skill requirements and higher levels. Currently the code scrapes the skill information from the official site where the skills are organized neatly in an HTML format. After which the code saves the formatted skills into a new file at a specific path. There are currently 2 issues regarding this including the code not adding the required skills such as different starting skills that the player receives upon the creation of the character as well as the lack of the capability to sacrifice 2 independent skills to get a skill of higher collective value that requires a intermediary skill to unlock.